PEAK-27 - Code Cleanup 1.0 / 1.1 - BIG - Core Logic Foundation REFACTORE + FIX + MIGRATION

+++ Application/Source/PeakPerformance.Application/Dtos/Users/Auth/SignupDto.cs
@@ -47,13 +47,13 @@ public class SignupDto
 
         user.UserRoles.Add(new UserRole
         {
-            RoleId = (int)eSystemRole.User
+            RoleId = eSystemRole.User
         });
 
         if (isSigningUp)
             user.UserActivityLogs.Add(new UserActivityLog
             {
-                ActionTypeId = (int)eActionType.Signup
+                ActionTypeId = eActionType.Signup
             });
     }
 }
